feat: add session layouts
