+++ src/app/(dashboard)/dashboard/page.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+    return (
+        <div>
+            <p>dashboard page</p>
+        </div>
+    )
+}
