+++ src/app/(session)/signin/signin.tsx
@@ -1,0 +1,7 @@
+export default function Page() {
+    return (
+        <div>
+            signin page
+        </div>
+    )
+}
